Split heavy alg.copy test
