+++ libcudacxx/test/libcudacxx/std/algorithms/alg.modifying/alg.copy/copy_fwd.pass.cpp
@@ -1,0 +1,41 @@
+//===----------------------------------------------------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+// <algorithm>
+
+// template<InputIterator InIter, OutputIterator<auto, InIter::reference> OutIter>
+//   constexpr OutIter   // constexpr after C++17
+//   copy(InIter first, InIter last, OutIter result);
+
+#include <cuda/std/__algorithm_>
+#include <cuda/std/cassert>
+
+#include "copy_common.h"
+
+TEST_CONSTEXPR_CXX20 __host__ __device__ bool test()
+{
+  test<forward_iterator<const int*>, cpp17_output_iterator<int*>>();
+  test<forward_iterator<const int*>, cpp17_input_iterator<int*>>();
+  test<forward_iterator<const int*>, forward_iterator<int*>>();
+  test<forward_iterator<const int*>, bidirectional_iterator<int*>>();
+  test<forward_iterator<const int*>, random_access_iterator<int*>>();
+  test<forward_iterator<const int*>, int*>();
+
+  return true;
+}
+
+int main(int, char**)
+{
+  test();
+
+#if TEST_STD_VER >= 2020
+  static_assert(test());
+#endif // TEST_STD_VER >= 2020
+
+  return 0;
+}
